Use trait for manager test
